:adhesive_bandage: (sonarcloud): Fix bugs and code smells
:adhesive_bandage: (otm8009a): Use std::byte for bitwise operation
:adhesive_bandage: (jpeg): Remove redundant cast
:adhesive_bandage: (dma2d): Fix implicit conversion loses
:adhesive_bandage: (jpeg): Add tests recommanded by sonarcloud
:adhesive_bandage: (hal): Remove unused acces specifier
:adhesive_bandage: (hal): Use =default as default implementation
:adhesive_bandage: const non-mutable member function
:adhesive_bandage: (ll): Remove unused namespace
:adhesive_bandage: (dsi): Avoid implicit conversion
:adhesive_bandage: (lcd) Add missing final keyword
:adhesive_bandage: (font): Add const in fontGetPixelBytes argument
:adhesive_bandage: (otm8009a): Replace "/*" by "//"
:adhesive_bandage: (otm8009a): Nested namespace

+++ drivers/LKCoreVideo/source/LKCoreDMA2D.cpp
@@ -49,8 +49,8 @@ void LKCoreDMA2D::initialize()
 	_hal.HAL_DMA2D_ConfigLayer(&_hdma2d, 1);
 }
 
-void LKCoreDMA2D::transferData(uintptr_t input_data_address, uintptr_t output_data_address, uint16_t width,
-							   uint16_t height)
+void LKCoreDMA2D::transferData(uintptr_t input_data_address, uintptr_t output_data_address, uint32_t width,
+							   uint32_t height)
 {
 	if (_hal.HAL_DMA2D_Init(&_hdma2d) == HAL_OK) {
 		if (_hal.HAL_DMA2D_ConfigLayer(&_hdma2d, 1) == HAL_OK) {
@@ -61,7 +61,7 @@ void LKCoreDMA2D::transferData(uintptr_t input_data_address, uintptr_t output_da
 	}
 }
 
-void LKCoreDMA2D::transferImage(uint16_t width, uint16_t height, uint16_t width_offset)
+void LKCoreDMA2D::transferImage(uint32_t width, uint32_t height, uint32_t width_offset)
 {
 	_hdma2d.Init.Mode				= DMA2D_M2M_PFC;
 	_hdma2d.LayerCfg[1].InputOffset = width_offset;
@@ -75,7 +75,7 @@ DMA2D_HandleTypeDef LKCoreDMA2D::getHandle(void)
 	return _hdma2d;
 }
 
-void LKCoreDMA2D::transferDrawing(uintptr_t first_pixel_address, uint16_t width, uint16_t height, uint32_t color)
+void LKCoreDMA2D::transferDrawing(uintptr_t first_pixel_address, uint32_t width, uint32_t height, uint32_t color)
 {
 	_hdma2d.Init.Mode		  = DMA2D_R2M;
 	_hdma2d.Init.OutputOffset = lcd::dimension.width - width;
